chore: split test_commands into multiple files
